+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * There's now a loading spinner while files load in the viewer
 * Highlight for current item in the list when editing a system user/group
 * Brand-spanking new file viewer using CodeMirror for themes etc
-* Databases can finally be listed on the Databases page
+* Databases can finally be created and listed on the Databases page
 * Current branches for a Site's repo are now listed in a dropdown
 
 ### Changed